Allow to define variables with hosts
When defining a variable type query in Grafana, it will call the metricFindQuery endpoint of your datasource, which is meant to match and return metrics. With this change, whenever the "query" of metricFindQuery is $hosts, it will request for hosts and return their names instead.

This change is meant to allow the definition of variables whose dynamic content are hosts (or metrics).